+++ tests/testthat/test-class-workbook-wrappers.R
@@ -47,7 +47,7 @@ test_that("wb_load() is a wrapper", {
   on.exit(unlink(tmp_xlsx), add = TRUE)
 
   wb_workbook()$add_worksheet()$add_data(x = iris)$save(tmp_xlsx)
-  expect_wrapper("load", params = list(file = tmp_xlsx), ignore_wb = TRUE)
+  expect_wrapper("load", params = list(file = tmp_xlsx), ignore_wb = TRUE, ignore_fields = "tmpDir")
 })
 
 # wb_save() ---------------------------------------------------------------